Mean shift with nice 3D presentation
This library is nice in its figure graphics, but selecting the data can be handled through a gitpoint() function in the matplotlib library. Presenting the point data itself is another story. I could not get the data to appear directly on the figure it was selected from, but I was able to create a second window to show the selected point.