Almost Production Ready
Only thing that does not work is the last loop in Blocker2.py. For some reason it is being killed by my computer with "Killed -9" after running the loop for about 7 minutes. The last loop should take nearly the same time as the .append(InData[k]) loop. Any help here is appropriated.